Remove is<TYPE> checking methods.
Previously we had implemented the methods "isConstraint",
"isIntVariable", and so on methods to provide the functionality
of verifying a given instance of the specified type was indeed of
that type. However, this is provided out-of-the-box by the `isa` method
which makes the previous code redundant and more verbose. Therefore,
this commit removes such placeholder code.

It also renames the constraint method that was previously called 'penalty'
to 'evaluate'.